Integra archive, transcendencia e XML Union
Atualiza o contexto tecnico do workspace e consolida as alteracoes feitas no cliente e no servidor durante a estabilizacao do Digimon Archive, transcendencia, D-Unit/XML Union, GM Panel, pacotes e Jump Booster. Este commit foi rebaseado sobre os commits remotos mais recentes da main antes do push.

Principais pontos:
- Documenta o fluxo correto de build Release_English, o uso obrigatorio de DProject.sln, os destinos de copia do executavel e o guia AGENTS.md/CodeGraph do workspace.
- Reorganiza a UI do Digimon Archive para seguir o comportamento oficial: cinco slots visiveis com setas de paginacao, suporte aos oito mercenarios locais e abas Storage/Academy/Transcendence.
- Estabiliza os pacotes do Archive 3201/3204, aceitando formatos com e sem byte VIP, validando slots, usando CurrentType no load, filtrando digimons arquivados em slot 255 e adicionando logs detalhados ARCHIVE 3201/3204.
- Corrige persistencia do Archive ao atualizar slots individuais e linhas de archive em vez de reordenar toda a lista de digimons, evitando perda visual/persistida apos abrir, fechar, trocar canal ou mover digimons repetidamente.
- Preserva a criacao defensiva de linha de slot do archive quando o banco nao possui a row esperada para um slot aberto, evitando falha silenciosa ao guardar digimon.
- Mantem a transcendencia compativel com o cliente English sem habilitar globalmente o define coreano, ajustando os layouts de carga de EXP 16039 e finalizacao 16040.
- Restaura o GM Panel moderno com a navegacao lateral, Make Item, Debug Char, Level Up, LevelUp-Digi, AllEvo e Bits, alem de corrigir o item/icone do main menu.
- Substitui a antiga entrada UnionDUnit no projeto pelo fluxo XmlUnionDUnit e amplia a janela XML Union/D-Unit com carregamento XML, cards, filtros, ranks, rewards, captura de input e cache de render.
- Adiciona suporte servidor para XML Union/D-Unit: loader de assets XML, servico de colecoes, DTO/model/config de progresso, migration, processors, packets de progresso/reward e XMLs runtime em dmo-server-main-postgres/xml/union.
- Audita e corrige divergencias de protocolo em login, cash shop, hot time, trade, party, quest, item socket, spirit craft, timeout de item, buff, escala, consigned shop, warehouse, guild, hatch, loot, initial info e pacotes relacionados.
- Corrige o crash do Jump Booster/All Area causado por message boxes pendentes: DelMsg nao cria dismissal sem caixa viva, DeleteList limpa pendencias e JumpBuster valida mapa/selecoes com fallback de envio direto.
- Ajusta loaders e queries de assets, mapas, mobs, drops, evolucoes, docker compose, schema validator e logs para facilitar diagnostico em futuras incompatibilidades de bin/pacote.
- Adiciona PROTOCOL_AUDIT_2026-05-22.md com o mapa das divergencias corrigidas e ignora Dcommerce/ por ser dump/tooling local grande, com XMLs extraidos, runtime e binarios que nao devem entrar como source.

Validacao feita antes do push:
- git diff --check sem erros apos o rebase.
- dotnet build C:\0-NewDmo\dmo-server-main-postgres\src\Source\Distribution\DigitalWorldOnline.Game.Host\DigitalWorldOnline.Game.csproj --no-restore /clp:ErrorsOnly: sucesso, 0 erros, 972 avisos existentes.
- MSBuild C:\0-NewDmo\dmo-client-main\DProject.sln /m /p:Configuration=Release_English /p:Platform=Win32 /v:minimal com CL=/FS: sucesso, gerando C:\Release\GSP\DProject_R_ENG.exe.
- Executavel copiado para C:\0-NewDmo\ClientDist\DProject_R_ENG.exe e D:\OpenDmo\src\Source\Cliente-OficialV2\Release\DProject_R_ENG.exe.
- SHA256 do executavel gerado: F280969FD7A52A511EE9FE5001B97B136CDE8DE2A74D61FA2E94D416E49AA8F7.

+++ dmo-server-main-postgres/src/Source/Distribution/DigitalWorldOnline.Game.Host/PacketProcessors/ItemSocketIdentifyPacketProcessor.cs
@@ -45,7 +45,7 @@ namespace DigitalWorldOnline.Game.PacketProcessors
             if (remaining < 6)
             {
                 client.Send(UtilitiesFunctions.GroupPackets(
-                    new ItemSocketIdentifyPacket(new ItemModel(), (int)client.Tamer.Inventory.Bits).Serialize(),
+                    new ItemSocketIdentifyPacket(new ItemModel(), client.Tamer.Inventory.Bits).Serialize(),
                     new LoadInventoryPacket(client.Tamer.Inventory, InventoryTypeEnum.Inventory).Serialize()));
                 return;
             }
@@ -96,7 +96,7 @@ namespace DigitalWorldOnline.Game.PacketProcessors
                 {
                     client.Send(
                         UtilitiesFunctions.GroupPackets(
-                            new ItemSocketIdentifyPacket(itemInfo, (int)client.Tamer.Inventory.Bits).Serialize(),
+                            new ItemSocketIdentifyPacket(itemInfo, client.Tamer.Inventory.Bits).Serialize(),
                             new LoadInventoryPacket(client.Tamer.Inventory, InventoryTypeEnum.Inventory).Serialize()
                         )
                     );
@@ -140,7 +140,7 @@ namespace DigitalWorldOnline.Game.PacketProcessors
                 await _sender.Send(new UpdateItemListBitsCommand(client.Tamer.Inventory));
                 client.Send(
                     UtilitiesFunctions.GroupPackets(
-                        new ItemSocketIdentifyPacket(itemInfo, (int)client.Tamer.Inventory.Bits).Serialize(),
+                        new ItemSocketIdentifyPacket(itemInfo, client.Tamer.Inventory.Bits).Serialize(),
                         new LoadInventoryPacket(client.Tamer.Inventory, InventoryTypeEnum.Inventory).Serialize()
                     )
                 );
@@ -150,7 +150,7 @@ namespace DigitalWorldOnline.Game.PacketProcessors
             {
                 client.Send(
                     UtilitiesFunctions.GroupPackets(
-                        new ItemSocketIdentifyPacket(new ItemModel(), (int)client.Tamer.Inventory.Bits).Serialize(),
+                        new ItemSocketIdentifyPacket(new ItemModel(), client.Tamer.Inventory.Bits).Serialize(),
                         new SystemMessagePacket($"Invalid item.").Serialize(),
                         new LoadInventoryPacket(client.Tamer.Inventory, InventoryTypeEnum.Inventory).Serialize()
                     )
